Update run_analysis.R
Using the "merge" function to join the Activity id to the Activity description was modifying the order of the Activity id, resulting in erroneous data. Using the "left_join" function from dplyr prevents this from happening. Also changed the reading of the activity labels file to not read strings as factors